refactor: strip out funcs from index into module/make func more reuseable
